Add a permanent store for some fallback ghosts
This addition is to deal with the question of how to place a ghost vault
if there are no bones files: basically, always try to keep around some
ghosts to use, rather than deleting all of them all the time. If this
idea works out, we may want to pre-seed download versions with some
sanitized ghosts.  This randomizes names saved in the permastore for
online play, but doesn't for offline play.