Add resumable_line_reader
This is intended for use as an interim measure, until I get around to
adding a text input widget. The base line_reader cannot be used in new
UI code, since the UI framework needs to provide the keys that arrive at
the widget. The resumable_line_reader works around that by being driven
externally with putkey().